+++ apps/backend/src/app/api/v1/users/crud.tsx
@@ -11,7 +11,7 @@ import { decodeBase64 } from "@stackframe/stack-shared/dist/utils/bytes";
 import { StackAssertionError, StatusError, captureError, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
 import { hashPassword } from "@stackframe/stack-shared/dist/utils/password";
 import { createLazyProxy } from "@stackframe/stack-shared/dist/utils/proxies";
-import { teamPrismaToCrud } from "../teams/crud";
+import { teamPrismaToCrud, teamsCrudHandlers } from "../teams/crud";
 
 export const userFullInclude = {
   projectUserOAuthAccounts: {
@@ -80,6 +80,7 @@ export const userPrismaToCrud = (prisma: Prisma.ProjectUserGetPayload<{ include:
     profile_image_url: prisma.profileImageUrl,
     signed_up_at_millis: prisma.createdAt.getTime(),
     client_metadata: prisma.clientMetadata,
+    client_read_only_metadata: prisma.clientReadOnlyMetadata,
     server_metadata: prisma.serverMetadata,
     has_password: !!prisma.passwordHash,
     auth_with_email: prisma.authWithEmail,
@@ -168,6 +169,7 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
         projectId: auth.project.id,
         displayName: data.display_name === undefined ? undefined : (data.display_name || null),
         clientMetadata: data.client_metadata === null ? Prisma.JsonNull : data.client_metadata,
+        clientReadOnlyMetadata: data.client_read_only_metadata === null ? Prisma.JsonNull : data.client_read_only_metadata,
         serverMetadata: data.server_metadata === null ? Prisma.JsonNull : data.server_metadata,
         primaryEmail: data.primary_email,
         primaryEmailVerified: data.primary_email_verified ?? false,
@@ -190,6 +192,24 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
     });
 
     const result = userPrismaToCrud(db);
+
+    if (auth.project.config.create_team_on_sign_up) {
+      await teamsCrudHandlers.adminCreate({
+        data: {
+          display_name: data.display_name ?
+            `${data.display_name}'s Team` :
+            data.primary_email ?
+              `${data.primary_email}'s Team` :
+              "Personal Team"
+        },
+        query: {
+          add_current_user: "true",
+        },
+        project: auth.project,
+        user: result,
+      });
+    }
+
 
     await sendUserCreatedWebhook({
       projectId: auth.project.id,
@@ -247,6 +267,7 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
         data: {
           displayName: data.display_name === undefined ? undefined : (data.display_name || null),
           clientMetadata: data.client_metadata === null ? Prisma.JsonNull : data.client_metadata,
+          clientReadOnlyMetadata: data.client_read_only_metadata === null ? Prisma.JsonNull : data.client_read_only_metadata,
           serverMetadata: data.server_metadata === null ? Prisma.JsonNull : data.server_metadata,
           primaryEmail: data.primary_email,
           primaryEmailVerified: data.primary_email_verified ?? (data.primary_email !== undefined ? false : undefined),