+++ apps/oauth-mock-server/src/index.ts
@@ -8,6 +8,7 @@ const mockedProviders = [
   "google",
   "microsoft",
   "spotify",
+  "discord"
 ];
 
 const configuration: Configuration  = {