+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,42 @@
 # demo-app
 
+## 2.5.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-ui@2.5.23
+  - @stackframe/stack@2.5.23
+  - @stackframe/stack-shared@2.5.23
+
+## 2.5.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.22
+  - @stackframe/stack-ui@2.5.22
+  - @stackframe/stack@2.5.22
+
+## 2.5.21
+
+### Patch Changes
+
+- Discord OAuth provider
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.21
+  - @stackframe/stack@2.5.21
+  - @stackframe/stack-ui@2.5.21
+
+## 2.5.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.20
+  - @stackframe/stack@2.5.20
+  - @stackframe/stack-ui@2.5.20
+
 ## 2.5.19
 
 ### Patch Changes