+++ examples/middleware/CHANGELOG.md
@@ -1,5 +1,33 @@
 # middleware-demo
 
+## 2.5.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.23
+
+## 2.5.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.22
+
+## 2.5.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.21
+
+## 2.5.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.5.20
+
 ## 2.5.19
 
 ### Patch Changes