+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @stackframe/stack-shared
 
+## 2.5.23
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.5.23
+
+## 2.5.22
+
+### Patch Changes
+
+- Team metadata
+  - @stackframe/stack-sc@2.5.22
+
+## 2.5.21
+
+### Patch Changes
+
+- Discord OAuth provider
+  - @stackframe/stack-sc@2.5.21
+
+## 2.5.20
+
+### Patch Changes
+
+- Improved account settings
+  - @stackframe/stack-sc@2.5.20
+
 ## 2.5.19
 
 ### Patch Changes